requirements.txt
# LLM & Orchestration
langchain>=0.3.0
langchain-ollama>=0.3.0
langgraph>=0.2.0
langsmith>=0.2.0

# Schema validation
pydantic>=2.0.0

# PDF
PyMuPDF>=1.24.0

# Config
pyyaml>=6.0

# Notebook
jupyter>=1.0.0
ipywidgets>=8.0.0
IPython>=8.0.0
